+++ kuhako/src/main/java/com/capstone/kuhako/services/ResellerServiceImpl.java
@@ -23,13 +23,13 @@ public class ResellerServiceImpl implements ResellerService {
     }
 
     // Delete reseller
-    public ResponseEntity deleteReseller(int id){
+    public ResponseEntity deleteReseller(Long id){
         resellerRepository.deleteById(id);
         return new ResponseEntity<>("User Deleted successfully", HttpStatus.OK);
     }
 
     // Update a reseller
-    public ResponseEntity updateReseller(int id, Reseller reseller ){
+    public ResponseEntity updateReseller(Long id , Reseller reseller ){
         // Find the reseller  to update
         Reseller resellerForUpdate = resellerRepository.findById(id).get();
         // Updating the reseller name and password
